Make hex_prefix a part of the merkle_patricia_tree
