--- apps/hex_prefix/.circleci/config.yml
@@ -1,46 +1,0 @@
-jobs:
-  build:
-    working_directory: ~/hex_prefix
-    docker:
-      - image: elixir:latest
-    steps:
-      - checkout
-      - restore_cache:
-          keys:
-            - v1-dependency-cache-{{ arch }}-{{ checksum "mix.lock" }}
-            - v1-dependency-cache-{{ arch }}
-            - v1-dependency-cache
-
-      - run: mix local.hex --force
-      - run: mix deps.get
-      - run: mix credo
-      - run: mix format --check-formatted
-      - run:
-          command: mix compile
-          environment:
-            MIX_ENV: test
-
-      - save_cache:
-          key: v1-dependency-cache-{{ arch }}-{{ checksum "mix.lock" }}
-          paths:
-            - _build
-            - deps
-            - ~/.mix
-
-      - run: mix test
-
-      - restore_cache:
-          keys:
-            - v1-plt-cache-{{ arch }}-{{ checksum "mix.lock" }}
-            - v1-plt-cache-{{ arch }}
-            - v1-plt-cache
-
-      - run: mix dialyzer --plt
-
-      - save_cache:
-          key: v1-plt-cache-{{ arch }}-{{ checksum "mix.lock" }}
-          paths:
-            - _build
-            - ~/.mix
-
-      - run: mix dialyzer --halt-exit-status
