--- apps/hex_prefix/.formatter.exs
@@ -1,3 +1,0 @@
-[
-  inputs: ["mix.exs", "{config,lib,test}/**/*.{ex,exs}"]
-]
