+++ apps/merkle_patricia_tree/mix.exs
@@ -48,7 +48,6 @@ defmodule MerklePatriciaTree.Mixfile do
       {:credo, "~> 0.9.1", only: [:dev, :test], runtime: false},
       {:poison, "~> 3.1.0", only: [:dev, :test], runtime: false},
       {:ex_doc, "~> 0.16", only: :dev, runtime: false},
-      {:hex_prefix, in_umbrella: true},
       {:ex_rlp, "~> 0.3.0"},
       {:keccakf1600, "~> 2.0.0", hex: :keccakf1600_orig},
       {:exleveldb, "~> 0.13.1"},
